+++ server/db/update_scripts/02_update_v4.9.0_v5.0.0.sql
@@ -379,3 +379,5 @@ END
 
 DELIMITER ;
 
+-- Add new style `actionBuilder`
+INSERT INTO `styles` (`name`, `id_type`, id_group, description) VALUES ('actionConfigBuilder', '1', (select id from styleGroup where `name` = 'intern' limit 1), 'Internal style used for the configuration of an action');