+++ CHANGELOG.md
@@ -8,8 +8,6 @@
  - add option for `keyword` change
  - add option for `page_access_type` change
  - add option for page `is_headless` change
- - add new style `conditionBuilder`
- - add new style `dataConfigBuilder`
 
 ### Bugfix 
  - fix json condition builder for `in_one_of` type used with other condition
@@ -19,6 +17,9 @@
 
 ### Internal
  - rework the `entryList` and `entryRecord` styles and how they work
+ - add new internal style `conditionBuilder`
+ - add new internal style `dataConfigBuilder`
+ - add new internal style `actionConfigBuilder`
 
 # v4.9.6
 ### New Features